creating temp files, moving code around, adding theme file
