+++ lib/src/config/constants.dart
@@ -1,0 +1,9 @@
+// This directory is for any constant information our app may use.
+// For example, we may repeat the appName in multiple spots. We create
+// a constant here, and just reference this everywhere so we do not have
+// to change the name in many spots.
+
+class AppConfig {
+  static const String appName = 'Food Group App';
+  // Other configurations
+}
