+++ lib/src/models/location.dart
@@ -1,0 +1,8 @@
+// This directory is for creating objects and 
+// basically listing all the characteristics of the model.
+// For example, a Person model may have a 
+// name, address, birth date, and more. We can then see
+// all attributes of a person and standardize that object in
+// our codebase
+
+// TODO create a location model
