+++ lib/src/routes/app_routes.dart
@@ -1,0 +1,7 @@
+// This folder is to declare all the routes / paths of our app
+
+class AppRoutes {
+  static const String home = '/';
+  static const String profile = '/locations';
+  // Other routes in the app
+}
