+++ lib/src/screens/locations/location_screen.dart
@@ -1,0 +1,1 @@
+// TODO should contain the add location screen code
